Extract SMS into its own class
