+++ lib/takeaway.rb
@@ -1,10 +1,11 @@
 require_relative 'menu'
+require_relative 'sms'
 require 'twilio-ruby'
 require 'date'
 class Takeaway
   # This class allows you to view and order dishes
 
-  attr_reader :items, :dishes, :total_order, :menu, :one_hour_ahead
+  attr_reader :items, :dishes, :total_order, :menu, :one_hour_ahead, :sms
 
   def initialize(menu = Menu.new)
     @menu = menu
@@ -28,25 +29,13 @@ class Takeaway
     fail_message = "Order total is £#{total_order_cost}. You expected £#{cost}."
     message = "Thank you! Your order costs £#{total_order_cost} and will be delivered before #{one_hour_ahead.strftime "%H:%M"}"
     fail fail_message if cost != total_order_cost
-    send_sms(message)
+    @sms = SMS.new
+    sms.send_sms(message)
     message
   end
 
   def one_hour_ahead
     DateTime.now + (1/24.0)
-  end
-
-  def send_sms(message)
-    account_sid = 'AC033ff5fd1ab50ecf84e702ae695c4053'
-    auth_token = '74465a4227537a433741dd0749629b3a'
-    client = Twilio::REST::Client.new account_sid, auth_token
-
-    client.messages.create(
-    from: ['+441256830268'],
-    to: ['+447515356421'],
-    body: message
-    )
-
   end
 
   private
